refactor: PostLinkInfo 패키지를 옮긴다
 - 기존 응답 타입에 맞추기 위해 ArchivePageController에서 반환 시
   PostLinkInfo를 문자열로 바꿔서 내보내고 있지만, 전용 타입인
   PostLinkInfo를 내보내도록 바꾸려고 한다.  이에 맞춰서 패키지를 옮긴다
