+++ src/main/java/org/gsh/genidxpage/service/WebPageParser.java
@@ -1,5 +1,6 @@
 package org.gsh.genidxpage.service;
 
+import org.gsh.genidxpage.web.response.PostLinkInfo;
 import org.jsoup.Jsoup;
 import org.jsoup.nodes.Document;
 import org.jsoup.nodes.Element;
@@ -36,27 +37,4 @@ public class WebPageParser {
         return result.toString();
     }
 
-    public class PostLinkInfo {
-
-        private final String baseUrl = "https://web.archive.org";
-        private String pageUrl;
-        private String pageTitle;
-
-        public PostLinkInfo(String pageUrl, String pageTitle) {
-            this.pageUrl = pageUrl;
-            this.pageTitle = pageTitle;
-        }
-
-        public String getPageUrl() {
-            return pageUrl;
-        }
-
-        public String getPageTitle() {
-            return pageTitle;
-        }
-
-        public String buildPageLink() {
-            return String.format("<a href=\"%s%s\">%s</a>", baseUrl, pageUrl, pageTitle);
-        }
-    }
 }
